+++ AdventOfCode2019/Day7/Day7.cs
@@ -10,9 +10,9 @@ namespace AdventOfCode2019
     {
         public static void ExecuteStarOne(string fileLocation = "Day7/Day7.txt")
         {
-            int[] programMemory = File.ReadAllText(fileLocation).Split(',').Select(int.Parse).ToArray();
+            long[] programMemory = File.ReadAllText(fileLocation).Split(',').Select(long.Parse).ToArray();
 
-            var inputs = Utilities.GetPermutations(new List<int>() { 0, 1, 2, 3, 4 }, 5);
+            var inputs = Utilities.GetPermutations(new List<long>() { 0, 1, 2, 3, 4 }, 5);
 
             int largestOutput = Int32.MinValue;
 
@@ -31,9 +31,9 @@ namespace AdventOfCode2019
 
         public static void ExecuteStarTwo(string fileLocation = "Day7/Day7.txt")
         {
-            int[] programMemory = File.ReadAllText(fileLocation).Split(',').Select(int.Parse).ToArray();
+            long[] programMemory = File.ReadAllText(fileLocation).Split(',').Select(long.Parse).ToArray();
 
-            var inputs = Utilities.GetPermutations(new List<int>() { 5, 6, 7, 8, 9 }, 5);
+            var inputs = Utilities.GetPermutations(new List<long>() { 5, 6, 7, 8, 9 }, 5);
 
             int largestOutput = Int32.MinValue;
 
@@ -53,23 +53,23 @@ namespace AdventOfCode2019
 
     public class ThrusterAmplifiers
     {
-        public int[] ProgramMemory { get; private set; }
+        public long[] ProgramMemory { get; private set; }
 
-        public int[] Phases { get; private set; }
+        public long[] Phases { get; private set; }
 
         public int Output { get; private set; }
 
         private IntCodeMachine a, b, c, d, e;
 
-        public ThrusterAmplifiers(int[] programMemory, int[] phases)
+        public ThrusterAmplifiers(long[] programMemory, long[] phases)
         {
             ProgramMemory = programMemory;
             Phases = phases;
-            a = new IntCodeMachine(programMemory, new List<int>() { phases[0] }.ToArray());
-            b = new IntCodeMachine(programMemory, new List<int>() { phases[1] }.ToArray());
-            c = new IntCodeMachine(programMemory, new List<int>() { phases[2] }.ToArray());
-            d = new IntCodeMachine(programMemory, new List<int>() { phases[3] }.ToArray());
-            e = new IntCodeMachine(programMemory, new List<int>() { phases[4] }.ToArray());
+            a = new IntCodeMachine(programMemory, new List<long>() { phases[0] }.ToArray());
+            b = new IntCodeMachine(programMemory, new List<long>() { phases[1] }.ToArray());
+            c = new IntCodeMachine(programMemory, new List<long>() { phases[2] }.ToArray());
+            d = new IntCodeMachine(programMemory, new List<long>() { phases[3] }.ToArray());
+            e = new IntCodeMachine(programMemory, new List<long>() { phases[4] }.ToArray());
         }
 
         public void Run()
