+++ AdventOfCode2019Test/Day5Tests.cs
@@ -15,9 +15,9 @@ namespace AdventOfCode2019Test
             string memory =
                 "3,21,1008,21,8,20,1005,20,22,107,8,21,20,1006,20,31,1106,0,36,98,0,0,1002,21,125,20,4,20,1105,1,46,104,999,1105,1,46,1101,1000,1,20,4,20,1105,1,46,98,99";
 
-            var input = new int[1] { 1 };
+            var input = new long[1] { 1 };
 
-            IntCodeMachine machine = new IntCodeMachine(Array.ConvertAll(memory.Split(','), int.Parse), input);
+            IntCodeMachine machine = new IntCodeMachine(Array.ConvertAll(memory.Split(','), long.Parse), input);
             machine.Run();
 
             Assert.AreEqual("999", machine.Output);
@@ -29,9 +29,9 @@ namespace AdventOfCode2019Test
             string memory =
                 "3,21,1008,21,8,20,1005,20,22,107,8,21,20,1006,20,31,1106,0,36,98,0,0,1002,21,125,20,4,20,1105,1,46,104,999,1105,1,46,1101,1000,1,20,4,20,1105,1,46,98,99";
 
-            var input = new int[1] { 8 };
+            var input = new long[1] { 8 };
 
-            IntCodeMachine machine = new IntCodeMachine(Array.ConvertAll(memory.Split(','), int.Parse), input);
+            IntCodeMachine machine = new IntCodeMachine(Array.ConvertAll(memory.Split(','), long.Parse), input);
             machine.Run();
 
             Assert.AreEqual("1000", machine.Output);
@@ -43,9 +43,9 @@ namespace AdventOfCode2019Test
             string memory =
                 "3,21,1008,21,8,20,1005,20,22,107,8,21,20,1006,20,31,1106,0,36,98,0,0,1002,21,125,20,4,20,1105,1,46,104,999,1105,1,46,1101,1000,1,20,4,20,1105,1,46,98,99";
 
-            var input = new int[1] { 9 };
+            var input = new long[1] { 9 };
 
-            IntCodeMachine machine = new IntCodeMachine(Array.ConvertAll(memory.Split(','), int.Parse), input);
+            IntCodeMachine machine = new IntCodeMachine(Array.ConvertAll(memory.Split(','), long.Parse), input);
             machine.Run();
 
             Assert.AreEqual("1001", machine.Output);
@@ -56,9 +56,9 @@ namespace AdventOfCode2019Test
         {
             string memory = "3,9,8,9,10,9,4,9,99,-1,8";
 
-            var input = new int[1] { 8 };
+            var input = new long[1] { 8 };
 
-            IntCodeMachine machine = new IntCodeMachine(Array.ConvertAll(memory.Split(','), int.Parse), input);
+            IntCodeMachine machine = new IntCodeMachine(Array.ConvertAll(memory.Split(','), long.Parse), input);
             machine.Run();
 
             Assert.AreEqual("1", machine.Output);
